% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Natural earth country data}
\format{An object of class \code{SpatialPolygonsDataFrame} with 255 rows and 65 columns.}
\usage{
countries
}
\description{
This dataset contains polygons of all the countries at a resolution of 10m.
}
\examples{
data(countries)
names(countries)

two.countries<-subset(countries,GEOUNIT \%in\% c("Sweden","Denmark")) #Get the polygon of Sweden and Denmark
plot(two.countries)

eu.countries<-subset(countries,CONTINENT \%in\% "Europe") #get the full european continent 
}
\references{
\url{http://www.naturalearthdata.com/}
}
\keyword{datasets}
