% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensembles_meta}
\alias{ensembles_meta}
\title{Metadata from the ensembles dataset}
\format{A data frame with 102 rows and 8 variables:
\describe{
 \item{STATION}{stations code}
 \item{STATION NAME}{Name of the station}
 \item{COUNTRY NAME}{Country were the station is located}
 \item{LAT}{Latitude in WGS84}
 \item{LON}{Longitude in WGS84}
 \item{ELEV}{Elevation of the station in meters}
 \item{START}{Start time of the measurements}
 \item{STOP}{Closing date of the station or latest measurement in the dataset}
}}
\usage{
ensembles_meta
}
\description{
Metadata from the ensembles dataset
}
\keyword{datasets}
