% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global_radiation}
\alias{global_radiation}
\title{Global Horizontal Radiation}
\format{An object of class \code{RasterStack} of dimension 180 x 360 x 13.}
\usage{
global_radiation
}
\description{
Dataset from nasa: \url{https://eosweb.larc.nasa.gov/cgi-bin/sse/global.cgi}
}
\details{
The monthly average amount of the total solar radiation incident on a horizontal 
surface at the surface of the earth for a given month, averaged for that month over 
the 22-year period (Jul 1983 - Jun 2005). Each monthly averaged value is evaluated 
as the numerical average of 3-hourly values for the given month.

Unit: kWh/m2/day 
Based on: Satellite data
}
\keyword{datasets}
