% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mask2_topo_1km}
\alias{mask2_topo_1km}
\title{Newest Land sea mask and topographic parameter}
\format{An object of class \code{RasterStack} of dimension 5520 x 8400 x 14.}
\usage{
mask2_topo_1km
}
\description{
GMTED2010 DEM for land/sea masking and topographic
parameters
}
\section{Slots}{

\describe{
\item{\code{tpi}}{?}

\item{\code{slope}}{slope with values between 0 and 1}

\item{\code{aspect}}{?}

\item{\code{Pvec}}{?}

\item{\code{Qvec}}{?}

\item{\code{alt}}{altitude averaged over 1km}

\item{\code{dist2coast}}{distance to the nearest ocean/sea}
}}

\keyword{datasets}
