% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sarah.grid}
\alias{sarah.grid}
\title{SARAH solar irradiance dataset}
\format{An object of class \code{RasterStack} of dimension 2601 x 2601 x 1.}
\usage{
sarah.grid
}
\description{
An example of the gridded SARAH dataset for 1997-02-02. The data is rasterized from the original netcdf files. 
The data projection is WGS84. The extent of the data is +/-65 latitude and +/-65 longitude.
}
\details{
The countries dataset can be used to subset SARAH for a specific region within these bounds.
}
\section{Slots}{

\describe{
\item{\code{Irr}}{Solar irradiance in W/m2}
}}

\examples{
library(raster)
data(countries)
data(sarah.grid)
data(ensembles_meta)
WGS84<-CRS("+init=epsg:4326")

obs.sweden<-ensembles_meta[which(ensembles_meta$`COUNTRY NAME`=="SWEDEN"),]
obs.sweden<-data.frame(obs.sweden)
coordinates(obs.sweden)<-~LON+LAT
proj4string(obs.sweden)<-WGS84

fun <- function() {
plot(countries,add=TRUE)
plot(obs.sweden,add=TRUE)
}

plot(sarah.grid,addfun=fun)

sweden<-subset(countries,GEOUNIT \%in\% c("Sweden"))
sarah.sub<-crop(sarah.grid,sweden)
sarah.sub<-mask(sarah.sub,sweden)



plot(sarah.sub,addfun=fun) #the sarah dataset misses the northern part of Sweden, not shown in the plot

#Example with the E-OBS mask: cropping sarah with the land/sea mask
data(mask2_topo_1km)
sarah_sea_mask<-crop(sarah.grid,mask2_topo_1km)
plot(sarah_sea_mask,addfun=fun)


}
\references{
\url{https://wui.cmsaf.eu/safira/action/viewDoiDetails?acronym=SARAH_V002}
}
\keyword{datasets}
