% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synops}
\alias{synops}
\title{Radiation data from the synoptic stations}
\format{A data frame with 353119 rows and 6 variables:
\describe{
 \item{STATION}{stations code similar to the STATION in *_qq_stations files}
 \item{DATEFIELD}{format YYYYMMDD}
 \item{PARAM}{field not in use}
 \item{VALUE}{y-coordinates of the station in the Dutch RD-coordinate system}
 \item{QC}{Globale Radiation in J/m^2 (W/m^2 for Zweden)}
 \item{SOURCE}{same as STATION because there is no blending for radiation.}
}}
\usage{
synops
}
\description{
Radiation data from the synoptic stations
}
\references{
\url{https://www.ecad.eu/}
}
\keyword{datasets}
