% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sarah.grid}
\alias{sarah.grid}
\title{SARAH solar irradiance dataset}
\format{An object of class \code{RasterStack} of dimension 2601 x 2601 x 1.}
\usage{
sarah.grid
}
\description{
An example of the gridded SARAH dataset for 1997-02-02. The data is rasterized from the original netcdf files. 
The data projection is WGS84.
}
\section{Slots}{

\describe{
\item{\code{Irr}}{Solar irradiance in W/m2}
}}

\references{
\url{https://wui.cmsaf.eu/safira/action/viewDoiDetails?acronym=SARAH_V002}
}
\keyword{datasets}
